app/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
requests==2.32.3
anyio==4.7.0

openai==1.59.6

google-cloud-firestore==2.20.0
google-api-python-client==2.154.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
